backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mongoengine
djongo
pymongo
python-decouple
Pillow
django-filter
geopy
requests
